server/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
requests>=2.32.0
python-dotenv>=1.0.0
numpy>=1.26.0
